requirements.txt
Flask==3.0.0
pandas==2.1.4
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
